requirements.txt
langgraph
langgraph-checkpoint-postgres
langchain
langchain-aws
langchain-community
langchain-core
boto3
pydantic
python-dotenv
langsmith
psycopg[binary,pool]
fastapi
uvicorn
colorlog
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
chromadb
pypdf
